requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.25.0
Django==3.1.7
django-admin-interface==0.15.6
django-cloudinary-storage==0.3.0
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-model-utils==4.1.1
django-solo==1.1.5
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
mercadopago==2.0.7
Pillow==8.2.0
PyMySQL==1.0.2
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.5
